moved simpler GUI event handlers out of main.cpp
